feat: 체크리스트 좋아요 삭제 기능 manageService 도입해 변경
