+++ backend/bang-ggood/src/main/java/com/bang_ggood/checklist/controller/ChecklistController.java
@@ -66,11 +66,4 @@ public class ChecklistController {
         checklistService.deleteChecklistById(user, id);
         return ResponseEntity.noContent().build();
     }
-
-    @DeleteMapping("/checklists/{id}/like")
-    public ResponseEntity<Void> deleteChecklistLikeByChecklistId(@AuthPrincipal User user,
-                                                                 @PathVariable("id") long id) {
-        checklistService.deleteChecklistLikeByChecklistId(user, id);
-        return ResponseEntity.noContent().build();
-    }
 }
