+++ backend/bang-ggood/src/main/java/com/bang_ggood/like/controller/ChecklistLikeController.java
@@ -4,6 +4,7 @@ import com.bang_ggood.auth.config.AuthPrincipal;
 import com.bang_ggood.like.service.ChecklistLikeManageService;
 import com.bang_ggood.user.domain.User;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -20,6 +21,14 @@ public class ChecklistLikeController {
     @PostMapping("/checklists/{id}/like")
     public ResponseEntity<Void> createChecklistLike(@AuthPrincipal User user, @PathVariable("id") Long id) {
         checklistLikeManageService.createLike(user, id);
+        return ResponseEntity.noContent().build();
+    }
+
+
+    @DeleteMapping("/checklists/{id}/like")
+    public ResponseEntity<Void> deleteChecklistLikeByChecklistId(@AuthPrincipal User user,
+                                                                 @PathVariable("id") long id) {
+        checklistLikeManageService.deleteLike(user, id);
         return ResponseEntity.noContent().build();
     }
 }
