+++ backend/bang-ggood/src/test/java/com/bang_ggood/like/controller/ChecklistLikeE2ETest.java
@@ -2,8 +2,14 @@ package com.bang_ggood.like.controller;
 
 import com.bang_ggood.AcceptanceTest;
 import com.bang_ggood.checklist.ChecklistFixture;
+import com.bang_ggood.checklist.domain.Checklist;
+import com.bang_ggood.checklist.repository.ChecklistRepository;
 import com.bang_ggood.checklist.service.ChecklistManageService;
+import com.bang_ggood.like.repository.ChecklistLikeRepository;
 import com.bang_ggood.like.service.ChecklistLikeManageService;
+import com.bang_ggood.room.RoomFixture;
+import com.bang_ggood.room.domain.Room;
+import com.bang_ggood.room.repository.RoomRepository;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import io.restassured.http.Header;
@@ -18,6 +24,12 @@ class ChecklistLikeE2ETest extends AcceptanceTest {
     private ChecklistManageService checklistManageService;
     @Autowired
     private ChecklistLikeManageService checklistLikeManageService;
+    @Autowired
+    private ChecklistRepository checklistRepository;
+    @Autowired
+    private RoomRepository roomRepository;
+    @Autowired
+    private ChecklistLikeRepository checklistLikeRepository;
 
     @DisplayName("체크리스트 좋아요 추가 성공")
     @Test
@@ -46,5 +58,21 @@ class ChecklistLikeE2ETest extends AcceptanceTest {
                 .when().post("/checklists/" + checklistId + "/like")
                 .then().log().all()
                 .statusCode(409);
+    }
+
+    @DisplayName("체크리스트 좋아요 삭제 성공")
+    @Test
+    void deleteChecklistLikeByChecklistId() {
+        Room room = roomRepository.save(RoomFixture.ROOM_1());
+        Checklist checklist = checklistRepository.save(
+                ChecklistFixture.CHECKLIST1_USER1(room, this.getAuthenticatedUser()));
+        checklistLikeRepository.save(ChecklistFixture.CHECKLIST1_LIKE(checklist));
+
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .header(new Header(HttpHeaders.COOKIE, this.responseCookie.toString()))
+                .when().delete("/checklists/" + checklist.getId() + "/like")
+                .then().log().all()
+                .statusCode(204);
     }
 }
